Fixed and updated error message for the following tests that no longer work due to semantic changes:
  flowanalysis/recursion/rec_obj-1
  flowanalysis/recursion/rec_obj-2
  flowanalysis/recursion/rec_obj-3

Allow casts to "non-types" because C(int) is both a type and an
expression.

Disable printing of candidate functions for 'this' function and
updated error message for a test due to following change:

Removed code to ignore class types as in

  var c : C;

It is now treated as

  move(c, nil:C)

Enabled inlining of functions that are not code generated because of
prelude/_chpl_base issue.

Updated a bunch of tests that referred to classes that did not declare
the types of all their fields without specifying a field type.  See
test/types/type_variables/deitz/test_clone_class1a.chpl as an example.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6451 3a8e244f-b0f2-452b-bcba-4c88e055c3ca