Implemented a first cut of an unoptimized form of dynamic dispatch on the 'this' argument to a method.  This will not work in the case where a generic type inherits from a class.
Added primitives SETCID and GETCID to access a dynamic id on class
types to determine what the dynamic type of a class is.

Implemented a coercion wrapper cache to avoid making multiple coercion
wrappers that are the same.  This uses the same cache code as
instantiation.

Added a test case to test dynamic dispatch.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6600 3a8e244f-b0f2-452b-bcba-4c88e055c3ca