Fixed a corner case in the coercion wrapper code to handle coercion wrappers on methods without parentheses.  These are only created when the 'this' argument is coerced which can only happen with the sync coercion.
Unified out of bounds error messages for sequences and arrays.

Simplified sequence indexing function and added support for indexing
into a sequence with a negative integer.  If sequence indexing started
at zero, we could generate the palindrome with scalar promotion on an
arithmetic sequence, s(-length..length), but it doesn't.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6674 3a8e244f-b0f2-452b-bcba-4c88e055c3ca