Unified the two mechanisms for instantiating the types of functions and classes used to handle the two cases:
  fun foo(x) { ... }

and

  fun foo(type t, x: t) { ... }

For better or worse, we now handle these the same way.  Added
Symbol::isTypeVariable to keep track of the case that a variable is
only taking a value in order to query its type as in the second case
above where the variable t is only used for its type.  Post-resolution
this code needs to be removed where possible.  It is a guess that for
array types we will need to keep the value since the domain is not
part of the type.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6680 3a8e244f-b0f2-452b-bcba-4c88e055c3ca