Now that we support automatic coercions from scalar types to strings, this checkin removes all fwrite() functions for scalar types, replacing them with a single fwrite() function that takes a string argument.  This has the benefits of (a) making sure that all file I/O is stringified identically to string casts (previously, we had different code for printing complexes and casting them to strings, e.g.); (b) making it easier to add new code to the fwrite() call -- error cases, locking, and such.
In making this change, I had to update a few of Shannon's file
I/O error tests, because the use of various coercions and
inlining were confusing the line number mechanism.  Since the
point of the tests was not to test line numbers or coercions,
but error cases on File I/O, I put in explicit coercions to
strings in order to avoid the wrappers and inlining that threw
off the line numbers.  Getting line numbers right is a known
issue, and a broader one than intended to be solved by this
checkin.

Also, updated the inlfunc1_*.good files as usual.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6955 3a8e244f-b0f2-452b-bcba-4c88e055c3ca