Fixed line numbers on setter and getter calls and updated the line numbers in 5 tests, making them correct.
Added keyword error used in a special error call, for example,

  error("illegal assignment to sequence's length");

This error message is a static error that is printed to the user if a
function that contains the error is called after parameter folding and
instantiation are done.  It greatly improves the error message in a
semi-ridiculous future test/types/seq/bradc/length.modify.chpl.
Updated this test's output and removed the future.  The filename and
line number associated with this error message is the filename and
line number of the call expression from which the function was called.

The error statement should only be used within parameter conditional
statements so that it is removed if the parameter conditional
evaluates to false.  If a function contains an error statement after
parameter folding and instantiation, it is a static error to call this
function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6967 3a8e244f-b0f2-452b-bcba-4c88e055c3ca