Updated Chapel's file implementation to report better and more portable error messages for writing to a file that is not open for writing.  This should get rid of the outstanding cygwin and sunos regressions.
Also restructured the file module so that a file's name,
path, and mode can't be changed once the file is open, but
this doesn't currently work due to the need to write a
setter for a class field with the same name (future filed
for this in previous test I committed).  The code to do this
is checked in here, but not enabled, and I've filed 3 futures
against this capability.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca