Added support for querying the width of a primitive type in a formal argument as in
  def foo(i: int(?k)) ...

which creates a foo that can take an int of any of the widths.  This
is done by replicating the function for each supported primitive width
thereby allowing coercions to fall out.

Used this new support to greatly simplify the modules that define
operators and functions on the primitive types.  This removes alot of
replicated code.

Changed the assignment to complex and float so that narrowing is not
allowed.  Following Java, you cannot write

  var f: float(32) = 2.0;

This requires a suffix or an explicit cast.  Apparently there is no
parallel to the parameter expression exception for floats.

Updated tests to obey this coercion rule.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6996 3a8e244f-b0f2-452b-bcba-4c88e055c3ca