Changed record allocation strategy to allocate records on the stack rather than the heap.
Added field FnSymbol::buildSetter to mean that the user specified that
this function is a 'var' function and requires a setter.  This used to
be done with the FnSymbol::retRef field.  The retRef field now means
that the function returns a reference.  It is set during function
resolution and possibly unset in cullOverReferences.

Started to add strides to parameter for loops, but did not finish this
work due to parsing problems and changing needs related to the rest of
the changes described here.

Added a PRIMITIVE_REF to indicate that a reference should point at
what another reference points at or at a value.  Nice simplifications
to PRIMITIVE_MOVE as a result of this addition.

Ordered records to codegen in an acceptable order to avoid a C error
of using an unspecified type.

Major improvements to cullOverReferences to remove references that do
not need to be references.

Fixed a bug in inlining regarding references and updated the output of
a complex numbers test that was not working correctly.

Expect regression of memory/shannon/printMemTable2.  Brad to fix.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7037 3a8e244f-b0f2-452b-bcba-4c88e055c3ca