Added support for scans.
Changed _min and _max functions that query the minimum and maximum
values of a type to min and max.  Changed the name of the builtin
reduction classes to _min and _max.  Updated numerous tests that use
these functions.

Added standard generic functions min and max that take two arguments
and that are implemented using the < and > comparator respectively.
Added wrapper min and max functions that take three or more arguments
and that are implemented in terms of the two argument functions.

  def min(x, y) return if x < y then x else y;
  def max(x, y) return if x > y then x else y;
  def min(x, y, z...?k) return min(min(x, y), (...z));
  def max(x, y, z...?k) return max(max(x, y), (...z));

The module version of these functions is decorated with the inline
pragma for an efficient implementation.

Moved constant checking after resolution.  This needs revamping.
Changed the output of two tests, making them legal, since the error,
writing a setter that assigns to the argument, is dead.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7072 3a8e244f-b0f2-452b-bcba-4c88e055c3ca