Fixed a few bugs:
Removed infinite loop in topological order calculation of records for
code generation that was seen when a class of a record of the same
class existed.

Fixed handling of record references so that culling of over-references
is more aggressive.  This is necessarily aggressive or we end up with
references to records that are on the stack.

Fixed debug code in normalize to use #IFDEF.

Enabled appending to sequences any subtype of the elt_type.

Set object default value to nil.

Changed sync record into a class and special cased the 'sync var'
pragma so that the sync class does not have a defaultValue set to nil,
but is rather constructed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca