Added a guard against instantiating a type or function with generic types.
Also moved BaseAST::pragmas vector to Symbol::pragmas as pragmas are
only used on symbols.  This also helps to make this clear, a pragma is
not a DefExpr or and ExprStmt, but rather a Symbol.  However, pragmas
can still be added to DefExprs or ExprStmts of DefExprs, and they are
automatically added to Symbol::pragmas.

This refactoring reduces the max heap size used on the Jacobi test
case after pass buildDefaultFunctions by 7M.  The max heap size by the
end of the program is mainly unchanged, growing even larger
(relatively) in functionResolution.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7088 3a8e244f-b0f2-452b-bcba-4c88e055c3ca