Added support for domain assignment that reallocates arrays declared over the assigned domain and preserves the data in these arrays.  This involved adding a new base class of all array classes called _abase. It relies on dynamic dispatch on generics.
The data is preserved by iterating over the intersection of the two
domains in the domain assignment.  Note that strided arrays and
domains are not yet supported.  Added a function _extended_euclid
which I copied from the ZPL runtime where it is cited as being from
Knuth Vol. 2, Section 4.5.2.  This function is required to compute the
intersection of strided domains.

Changed arrays from records to classes.  This happens to remove the
duplicate data that we were allocating for arrays.  Updated the two
memory test outputs accordingly.

Fixed a bug in cleanup in which use statements were not cleaned up
early enough.

Improved how function calls are printed out so that if the actual is a
parameter, the value is written rather than its type.  This is nice if
the function expects a parameter though not ideal otherwise.  Also
improved how functions are printed out, now printing param in front of
arguments that have parameter intents.  Updated error outputs on about
9 tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7150 3a8e244f-b0f2-452b-bcba-4c88e055c3ca