Added initial support for constructors.  You can now write your own constructors (it is a constructor because it has the same name as the class) but there is no way to call another constructor, etc.  Added a test case to show it works.
Added a couple of error-catching constructors to the arithmetic
sequence class that use the error construct.  These constructors catch
the case where the user is trying to construct an arithmetic sequence
out of non-integers.

Removed trivial/bradc/uintloop.future and updated .good to admit that
this is correctly an error.  Brad- take a look.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7188 3a8e244f-b0f2-452b-bcba-4c88e055c3ca