This checkin gets the serial version of the NAS FT benchmark in Chapel up and running!  This is a milestone but note that it runs very slowly.
Temporarily disabled folding of floating point parameters and fixed
the parsing of complex literals to operate on the character string
like the parsing of float literals.

Fixed a bug in the working version of the NAS FT serial Chapel code in
which I had written "m-1" instead of "m-l".

Updated the NAS FT serial Chapel code to use scientific notation for
the floating point literals.

Removed the NOTEST file in test/npb/ft.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7198 3a8e244f-b0f2-452b-bcba-4c88e055c3ca