Added an error message test written by Mary Beth, now working.
Fixed a bug in which an expression of any class type could be assigned
to a variable of any class type.  The rhs now must be a subclass of
the lhs.

Fixed two buggy tests.  One tried to assign one variable to another
variable of subclass type.  I put in an explicit cast.  (Should this
cast be checked at runtime?  I believe yes.)

The other buggy test had written
  var d: domain(1) = [1..n, 1..n];
and had worked because the 2D domain is stored in memory such that it
could be accessed as if it were a 1D domain.  I fixed this test by
dropping the second range from the domain.

Special-cased assignment of nil to non-class error.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7206 3a8e244f-b0f2-452b-bcba-4c88e055c3ca