Fixed regression: [Error matching compiler output for classes/deitz/test_type_in_class2]
Special-cased the use of a class as a type where the class is generic
but has a default on the type.  See the test for an example.  This
special case is required because of not distinguishing between typedef
and type in classes.  It was expected and is easy to handle.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7277 3a8e244f-b0f2-452b-bcba-4c88e055c3ca