Adding a Chapel and a C version of the Random Access HPC benchmark.
I started with the C version, extracting it from the HPC benchmark
suite.  I had to modify it in order to make it a standalone program.

Then, I did a pretty straightforward translation from C to Chapel.
It's gross.  But, I wanted to make sure that the random number generation
matched the C version.  Much more work is needed still.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7294 3a8e244f-b0f2-452b-bcba-4c88e055c3ca