Changed the way in which ast wrapper nodes are automatically inserted. They are no longer (unsafely) added based on the dynamic type of a node in alist code.  They are now added when the list is accessed via a BaseAST.  So for example
  fn->insertAtTail(new DefExpr(var));

will put an ExprStmt around the DefExpr, but

  fn->body->body->insertAtTail(new DefExpr(var));

will not (resulting in a problem).

Wrappers that are built are ExprStmt on expressions when inserted into
a list that contains statements, DefExpr on ArgSymbols when inserted
as a formal (use insertFormalAtTail or insertFormalAtHead now), and
SymExpr on symbols when inserted into a CallExpr.

An ExprStmt is also put around expressions that are inserted via
insertBefore or insertAfter in a list of statements.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7330 3a8e244f-b0f2-452b-bcba-4c88e055c3ca