Changed the way error messages are reported to look more like gcc.
Added "In function 'foo':" to error messages if we know what function
the error occurred in.

Reworked check_normalized and put a catch to only show the error for a
variable undeclared to only print it once.  This gets it down to twice
for

  test/studies/hpcc/RA/bradc/ra-bradc-duplicateErrors.chpl

because of nested functions being pulled out.

Added a catch so the default == and != functions for records are not
built if one already exists.

Updated .goods.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7351 3a8e244f-b0f2-452b-bcba-4c88e055c3ca