Initial checkin of new spec with section The Conditional Statement created.
Stubs for new sections are automatically created by a preprocessing
perl script.  This also checks for duplicate section names and for
unused files (changed names).

A good start to the table of contents is complete for all but two
sections:  Reductions and Scans, and Input and Output.

We may want a more detail subsection breakdown for the Language
Overview section as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7359 3a8e244f-b0f2-452b-bcba-4c88e055c3ca