Add NOOP primitive for use in the implementation.
Added CallExpr* FnSymbol::visiblePoint to point to a spot in the code
that a function should look at to determine what functions are
visible.  This is particularly important for functions with class
arguments which are globally visible (and so denested before function
resolution).  This will also be used to mark the point of
instantiation of a function.

Stopped treating all functions as globally visible.

Specialized global function determination so that functions that take
arguments of type of the array or domain class are not treated
globally.

Changed "In function" message to print out outermost function for
nested functions.

This fixes regression of functions/bradc/intents-error.

Removed .future for studies/hpcc/RA/bradc/ra-bradc-duplicateErrors.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7377 3a8e244f-b0f2-452b-bcba-4c88e055c3ca