Expect 1 regression:
[Error matching compiler output for studies/hpcc/RA/bradc/ra-bradc-old1]

Changed all builtin operators on integers and unsigned integers so
that they operate solely on 32- and 64-bit integers and unsigned
integers.  So for

  var i: int(16);
  i = i + 1;

you need a cast

  var i: int(16);
  i = (i+1):int(16);

to show the potential data loss.  The following operators are changed:
  + - ~ !  + - * / ** % & | ^ == != <= >= < > && || << >>

Fixed << and >> operators to take their type from the left operand.
The right operand must be an integral type (following C#).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7496 3a8e244f-b0f2-452b-bcba-4c88e055c3ca