Enabled complex cast on a literal tuple so that something like
  (2, 3):complex

is a complex number.  It is currently not folded (but should be) so it
cannot be used in a parameter specification.

A literal tuple is required so

  var t = (2, 3);
  ... t:complex ...

would result in an error.  Is this a wrong choice?

Removed internal function _complex and updated lots of instances in
the test suite where this internal function was being used.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7517 3a8e244f-b0f2-452b-bcba-4c88e055c3ca