Added test case to show that strided slices of multidimensional arrays is also working.  But note that currently
  A(1..6, 1..6)

is a 2D slice that is identical to

  A[1..6, 1..6]

Since this is slicing, it does not use promotion which would have a
different result.

This seems wrong.  I believe A(1..6, 1..6) should refer to the
diagonal but I can see arguments that it shouldn't.  There are
diffculties with making it the diagonal that we should discuss.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7525 3a8e244f-b0f2-452b-bcba-4c88e055c3ca