Added support for folding of primitive casts that are applied to parameters.
Note: casts of parameters to type complex and string are not folded.

Reconstituted test/studies/hpcc/RA/bradc/ra-bradc-badshift to declare
parameters instead of variables in two cases.

Fixed a bug in which parameters that were used in the then or else
expression of a conditional expression would not be folded.  Removed
.future for test/studies/hpcc/RA/bradc/ra-bradc-badparam.

Fixed parameter coercion to only work on 32-bit integer parameters,
necessary now that casts are folded.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7530 3a8e244f-b0f2-452b-bcba-4c88e055c3ca