Note: regression: (see reason below)  [Error matching compiler output for arrays/bradc/stridedArrayArg2]
Added special handling of -= to allow it to call the domain's remove
method.

Made op= insert a cast to handle the case:

  var i: int(8) = 2;
  i += 1;

The issue here is that we cannot translate it into

  i = i + 1;

since this would result in an error since i + 1 is a 32-bit int.  Thus
op= is now handled as

  a op= b

is changed into

  a = (a op b):a.type

following C# semantics.

Unfortunately this breaks support for op= on whole arrays since we
don't want to cast it to the array.  This will require major reworking
of casts to get right.  This is the reason for the regression.

Implemented a first pass on inference on return types to infer the
return type to be any returned type that other returned values can be
coerced into including via the implicit coercion rule so

  return i:int(8);
  return 2;

will return an int(8) since 2 can be cast into it.  We discussed a
more restrictive rule at the meeting but I had trouble figuring out
the semantics of what we wanted.  Treat this as preliminary.

Added a future on domain remove that is the same as the last one but
uses -= instead of the method call.

Removed conditional expression hpcc future.

Added two tests to look at return inference and coercion.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7541 3a8e244f-b0f2-452b-bcba-4c88e055c3ca