Checking in the latest versions of ra2.chpl, ra2.good, ra3.chpl, ra3.good. These are getting very close to final versions...
Both versions of Random Access now have output indicating the number of
random streams and the length of those streams.  The ra3.chpl version
also prints out the length of the inner vector loop.

Both now print out a GUPs number and timing information.  The timer
routines are still not implemented, however.

The ra2 version is the Chapel version of the MTA code.  The ra3 version
is the ra2 version + an inner vector loop.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7558 3a8e244f-b0f2-452b-bcba-4c88e055c3ca