Made arithmetic domains generic over integral types so that in
  [1..10:uint, 1..10:uint]

the "dimensional index type" is uint.

Note that all arithmetic sequences must have the same element type and
this element type becomes the dimensional index type.  I added a
future to wait on a better error message if this is not the case.

The dimensional index type can be specified in a domain's type by
writing, for example, domain(2, uint).

Note: a dimensional index type of uint(64) is not YET working.

Made abs work over int, int(64), uint, and uint(64) to avoid getting a
float when passing it a uint.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7580 3a8e244f-b0f2-452b-bcba-4c88e055c3ca