Updated ra-bradc.chpl to utilize several new features:
        - uint domains
        - better shift operators
        - inout arguments
        - the bits() function
        - timers
        - [i in D]-style array declarations

and moved the last version to ra-bradc-old2.chpl.

This is starting to look like really great code to me!
(I haven't looked at Mary Beth's version in quite some
time because it's been easier to understand improvements
in the compiler in the code I'm familiar with;  need
to synch back up with where Mary Beth's code is soon).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7608 3a8e244f-b0f2-452b-bcba-4c88e055c3ca