Checking in a version of STREAMS that initializes the vectors with random numbers.  **This is different from the provided reference implementation.**
This version uses the Random module which is still under development.

The routine that checks the results needed to be modified to account for
the vectors being initialized with random numbers rather a constant value
for all entries.  I used the l_1-norm to compute the relative error in the
computed solution.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7646 3a8e244f-b0f2-452b-bcba-4c88e055c3ca